real_estate_broker_ai_agent/requirements.txt
fastapi
uvicorn
spacy
pydantic
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
sqlalchemy
